Split sim.rs to node.rs and net.rs
