+++ src/net.rs
@@ -26,255 +26,15 @@
 // For now, because lots of stuff isn't implemented yet:
 #![allow(dead_code)]
 
-use std::cmp::{Ordering, min};
 use std::mem;
-use std::hash::{Hash, Hasher};
-use std::fmt::{self, Formatter, Binary, Debug};
 use std::collections::HashMap;
 use std::collections::hash_map::Entry;
-use std::u64;
 
 use rand::{thread_rng, Rng};
-use rand::distributions::{Range, IndependentSample};
 
 use {NN, RR, Error, Result};
 use attack::AttackStrategy;
-
-
-// In the future, we may be able to do this:
-// const RANGE_NN: Range<NN> = Range::new(0, NN::max_value());
-#[allow(non_snake_case)]
-fn sample_NN() -> NN {
-    thread_rng().gen()
-}
-fn sample_ub(ub: NN) -> NN {
-    let range = Range::new(0, ub);
-    range.ind_sample(&mut thread_rng())
-}
-
-
-/// Name type (Xorable in routing library).
-trait NameT: Ord {
-    /// Returns the length of the common prefix with the `other` name; e. g.
-    /// the when `other = 11110000` and `self = 11111111` this is 4.
-    fn common_prefix(&self, other: Self) -> usize;
-
-    /// Compares the distance of the arguments to `self`. Returns `Less` if `lhs` is closer,
-    /// `Greater` if `rhs` is closer, and `Equal` if `lhs == rhs`. (The XOR distance can only be
-    /// equal if the arguments ar equal.)
-    fn cmp_distance(&self, lhs: Self, rhs: Self) -> Ordering;
-
-    /// Returns `true` if the `i`-th bit is `1`.
-    fn bit(&self, i: usize) -> bool;
-
-    /// Returns a copy of `self`, with the `index`-th bit set to `bit`.
-    ///
-    /// If `index` exceeds the number of bits in `self`, an unmodified copy of `self` is returned.
-    fn with_bit(self, i: usize, bit: bool) -> Self;
-
-    /// Returns a binary format string, with leading zero bits included.
-    fn binary(&self) -> String;
-
-    /// Returns a copy of self with first `n` bits preserved, and remaining bits
-    /// set to 0 (val == false) or 1 (val == true).
-    fn set_remaining(self, n: usize, val: bool) -> Self;
-}
-
-impl NameT for NN {
-    fn common_prefix(&self, other: Self) -> usize {
-        (self ^ other).leading_zeros() as usize
-    }
-
-    fn cmp_distance(&self, lhs: Self, rhs: Self) -> Ordering {
-        Ord::cmp(&(lhs ^ self), &(rhs ^ self))
-    }
-
-    fn bit(&self, i: usize) -> bool {
-        let pow_i = 1 << (mem::size_of::<Self>() * 8 - 1 - i); // 1 on bit i.
-        self & pow_i != 0
-    }
-
-    fn with_bit(mut self, i: usize, bit: bool) -> Self {
-        if i >= mem::size_of::<Self>() * 8 {
-            return self;
-        }
-        let pow_i = 1 << (mem::size_of::<Self>() * 8 - 1 - i); // 1 on bit i.
-        if bit {
-            self |= pow_i;
-        } else {
-            self &= !pow_i;
-        }
-        self
-    }
-
-    fn binary(&self) -> String {
-        format!("{1:00$b}", mem::size_of::<Self>() * 8, self)
-    }
-
-    fn set_remaining(self, n: usize, val: bool) -> Self {
-        let bits = mem::size_of::<NN>() * 8;
-        if n >= bits {
-            self
-        } else {
-            let mask = !0 >> n;
-            if val { self | mask } else { self & !mask }
-        }
-    }
-}
-
-
-// A group prefix, i.e. a sequence of bits specifying the part of the network's name space
-// consisting of all names that start with this sequence.
-#[derive(Clone, Copy, Default, Eq, Ord)]
-pub struct Prefix {
-    bit_count: usize,
-    name: NN,
-}
-
-impl Prefix {
-    /// Creates a new `Prefix` with the first `bit_count` bits of `name`.
-    /// Insignificant bits are all set to 0.
-    fn new(bit_count: usize, name: NN) -> Prefix {
-        Prefix {
-            bit_count: bit_count,
-            name: name.set_remaining(bit_count, false),
-        }
-    }
-
-    /// Returns `self` with an appended bit: `0` if `bit` is `false`, and `1` if `bit` is `true`.
-    fn pushed(mut self, bit: bool) -> Prefix {
-        self.name = self.name.with_bit(self.bit_count, bit);
-        self.bit_count += 1;
-        self
-    }
-
-    /// Returns a prefix copying the first `bitcount() - 1` bits from `self`,
-    /// or `self` if it is already empty.
-    fn popped(mut self) -> Prefix {
-        if self.bit_count > 0 {
-            self.bit_count -= 1;
-            // unused bits should be zero:
-            self.name = self.name.with_bit(self.bit_count, false);
-        }
-        self
-    }
-
-    /// Returns the number of bits in the prefix.
-    fn bit_count(&self) -> usize {
-        self.bit_count
-    }
-
-    /// Returns `true` if `self` is a prefix of `other` or vice versa.
-    fn is_compatible(&self, other: Prefix) -> bool {
-        let i = self.name.common_prefix(other.name);
-        i >= self.bit_count || i >= other.bit_count
-    }
-
-    fn common_prefix(&self, name: NN) -> usize {
-        min(self.bit_count, self.name.common_prefix(name))
-    }
-
-    /// Returns `true` if this is a prefix of the given `name`.
-    fn matches(&self, name: NN) -> bool {
-        self.name.common_prefix(name) >= self.bit_count
-    }
-}
-
-impl PartialEq<Prefix> for Prefix {
-    fn eq(&self, other: &Self) -> bool {
-        self.is_compatible(*other) && self.bit_count == other.bit_count
-    }
-}
-
-impl PartialOrd<Prefix> for Prefix {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        if self == other {
-            Some(Ordering::Equal)
-        } else if self.is_compatible(*other) {
-            None
-        } else {
-            Some(self.name.cmp(&other.name))
-        }
-    }
-}
-
-impl Hash for Prefix {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        for i in 0..self.bit_count {
-            self.name.bit(i).hash(state);
-        }
-    }
-}
-
-impl Binary for Prefix {
-    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
-        let mut binary = self.name.binary();
-        binary.truncate(self.bit_count);
-        write!(formatter, "Prefix({})", binary)
-    }
-}
-
-impl Debug for Prefix {
-    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
-        Binary::fmt(self, formatter)
-    }
-}
-
-
-/// Type of a node name
-pub type NodeName = u64;
-
-/// Generate a new node name
-pub fn new_node_name() -> NodeName {
-    sample_NN()
-}
-
-/// Data stored for a node
-#[derive(Clone, Copy)]
-pub struct NodeData {
-    age: u32, // initial age is 0
-    churns: u32, // initial churns is 0
-    is_malicious: bool,
-}
-
-impl NodeData {
-    /// New data (initial age and churns, not malicious)
-    pub fn new() -> Self {
-        NodeData {
-            age: 0,
-            churns: 0,
-            is_malicious: false,
-        }
-    }
-
-    /// New data (initial age and churns, is malicious)
-    pub fn new_malicious() -> Self {
-        NodeData {
-            age: 0,
-            churns: 0,
-            is_malicious: true,
-        }
-    }
-
-    /// Get the age
-    pub fn age(&self) -> u32 {
-        self.age
-    }
-
-    // Increment churns, and return whether this is high enough for relocation
-    fn churn_and_can_age(&mut self) -> bool {
-        self.churns += 1;
-        self.churns >= 2u32.pow(self.age)
-    }
-
-    /// Is this node malicous?
-    pub fn is_malicious(&self) -> bool {
-        self.is_malicious
-    }
-}
-
-/// Type of a node
-pub type Node = (NodeName, NodeData);
+use node::{Prefix, NodeName, NodeData};
 
 
 pub trait AddRestriction {
@@ -291,11 +51,11 @@ impl AddRestriction for NoAddRestriction {}
 pub struct RestrictOnePerAge;
 impl AddRestriction for RestrictOnePerAge {
     fn can_add(node_data: &NodeData, group: &HashMap<NodeName, NodeData>) -> bool {
-        let age = node_data.age;
+        let age = node_data.age();
         if age > 1 {
             return true;
         }
-        group.values().filter(|data| data.age == age).count() < 2
+        group.values().filter(|data| data.age() == age).count() < 2
     }
 }
 
@@ -363,6 +123,8 @@ impl Network {
     /// Return the number of nodes dropped.
     pub fn probabilistic_drop(&mut self, p: RR) -> usize {
         let thresh = (p * (NN::max_value() as RR)).round() as NN;
+        #[allow(non_snake_case)]
+        let sample_NN = || -> NN { thread_rng().gen() };
         let mut need_merge = vec![];
         let mut num = 0;
         for (prefix, ref mut group) in &mut self.groups {
@@ -379,7 +141,7 @@ impl Network {
 
         // Do any merges needed (after all removals)
         while let Some(prefix) = need_merge.pop() {
-            if prefix.bit_count == 0 {
+            if prefix.bit_count() == 0 {
                 // Not enough members in network yet; nothing we can do
                 continue;
             }
@@ -459,12 +221,12 @@ impl Network {
         let (group0, group1): (Group, Group) = old_group.into_iter()
             .partition(|node| prefix0.matches(node.0));
         for (name, data) in &group0 {
-            if data.is_malicious {
+            if data.is_malicious() {
                 attack.split(prefix, prefix0, *name, data);
             }
         }
         for (name, data) in &group1 {
-            if data.is_malicious {
+            if data.is_malicious() {
                 attack.split(prefix, prefix1, *name, data);
             }
         }
@@ -496,8 +258,8 @@ impl Network {
                 continue;   // skip this node
             }
             if node_data.churn_and_can_age() {
-                if to_relocate.map_or(true, |n| node_data.churns > n.1) {
-                    to_relocate = Some((*node_name, node_data.churns));
+                if to_relocate.map_or(true, |n| node_data.churns() > n.1) {
+                    to_relocate = Some((*node_name, node_data.churns()));
                 }
             }
         }
@@ -509,16 +271,16 @@ impl Network {
         if group.len() <= self.min_group_size {
             // Relocation is blocked to prevent the group from becoming too small,
             // but we still need the node to age.
-            group.get_mut(&to_relocate).expect("have node").age += 1;
+            group.get_mut(&to_relocate).expect("have node").incr_age();
             return None;
         }
 
         // Remove node, age and return:
         let mut node_data = group.remove(&to_relocate).expect("have node");
-        node_data.age += 1;
+        node_data.incr_age();
         trace!("Relocating a node with age {} and churns {}",
-               node_data.age,
-               node_data.churns);
+               node_data.age(),
+               node_data.churns());
         Some((to_relocate, node_data))
     }
